@sayedbilalbari 2025-07-30
Moving to draft as this is currently blocked on #1702 as while passing a long list of eventlogs, this results in the error - `OSError: [Errno 7] Argument list too long: '/bin/bash'` due to the size of the final command string rendered and the OS limitation mentioned above in the caveats. Will revisit this post working on the above issue.