@sayedbilalbari 2025-08-19
> The changes in the build and tests are not light changes. The `build.sh` used to take arguments then pass it along while building the jars/wheel-packages. In this PR, this is not the case anymore. The `tox.ini` calls the `build.sh` which is a major flow change to be sneaked-in with another PR With these changes, where should those variables be set? Are they still being passed to the mvn command? Later, when we want to set out CI/CD to have different build/tests what is the new way to achieve this?

Thanks @amahussein . Just some clarifying points.
- Nothing has changed in the original functionality of the build script. It remains the same with an added functionality to skip buliding the whl file. Any existing usage of build script remains the same.
- Tox setup previously relied on the trigger of the build script to install python package, still remains the case. We just moved away from doing that via a setup_env.sh script to doing that as part of a tox step. We still are still setting those mvn args in the tox script by means of env variables. [Code](https://github.com/NVIDIA/spark-rapids-tools/blob/fbdfb7aa8e0bde5e60ea4aa746ae8cdf89bb1dcd/user_tools/tox.ini#L111-L114) as below -
```
setenv =
    TOOLS_BUILDVER = {env:TOOLS_BUILDVER:350}
    TOOLS_HADOOP_VERSION = {env:TOOLS_HADOOP_VERSION:3.3.6}
    TOOLS_MAVEN_ARGS = -Dbuildver={env:TOOLS_BUILDVER} -Dhadoop.version={env:TOOLS_HADOOP_VERSION}
```
- Since this step still values ENV variables over the values set in the script, all we have to do to override is set env variables before triggering the tox and it will work as expected.
- If we want to change this, just update the defaults in the script or resort to setting these as env variables for different CI/CD and test runs