@sayedbilalbari 2025-08-19
Updated the TOOLS_MAVEN_ARGS to also be set by means of an environment variable